chore: move client.py to clients/http.py
