+++ surrealdb/clients/__init__.py
@@ -13,11 +13,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from .clients import SurrealDBHTTPClient
+from .http import SurrealDBHTTPClient
 
-__version__ = "0.1.0"
-__title__ = "surrealdb"
-__author__ = "SurrealDB"
-__license__ = "Apache-2.0"
-__copyright__ = "SurrealDB Ltd"
 __all__ = ("SurrealDBHTTPClient",)
